# Optimize BindingKey string creation.