--- a/src/Benchmarks/BenchmarkRunner/Benchmarks.cs
+++ b/src/Benchmarks/BenchmarkRunner/Benchmarks.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using ExRam.Gremlinq.Core;
@@ -21,7 +22,7 @@ namespace Benchmarks
         }
 
 
-        [Benchmark]
+        //[Benchmark]
         public async Task LargeGraphson3Path()
         {
             await _g
@@ -29,6 +30,46 @@ namespace Benchmarks
                 .V<Person>()
                 .Cast<ExRam.Gremlinq.Core.GraphElements.Path[]>()
                 .ToArrayAsync();
+        }
+
+        [Benchmark]
+        public void Old()
+        {
+            var key = 5326256;
+            var stringKey = string.Empty;
+
+            do
+            {
+                stringKey = (char)('a' + key % 26) + stringKey;
+                key /= 26;
+            } while (key > 0);
+
+            stringKey = "_" + stringKey;
+        }
+
+        [Benchmark]
+        public void New()
+        {
+            var key = 5326256;
+            var digits = key > 0
+                ? (int)Math.Ceiling(Math.Log(key + 1, 26)) + 1
+                : 2;
+
+            string.Create(
+                digits,
+                (key, digits),
+                (span, tuple) =>
+                {
+                    var (key, digits) = tuple;
+
+                    span[0] = '_';
+
+                    for (var i = digits - 1; i >= 1; i--)
+                    {
+                        span[i] = (char)('a' + key % 26);
+                        key /= 26;
+                    }
+                });
         }
 
         private static string GetJson(string name)