--- a/src/ExRam.Gremlinq.Core/Serialization/BindingKey.cs
+++ b/src/ExRam.Gremlinq.Core/Serialization/BindingKey.cs
@@ -16,6 +16,7 @@ namespace ExRam.Gremlinq.Core
                 _stringKey = Keys[key]._stringKey;
             else
             {
+                #if NETSTANDARD2_0
                 var stringKey = string.Empty;
 
                 do
@@ -25,6 +26,27 @@ namespace ExRam.Gremlinq.Core
                 } while (key > 0);
 
                 _stringKey = "_" + stringKey;
+                #else
+                var digits = key > 0
+                    ? (int)Math.Ceiling(Math.Log(key + 1, 26)) + 1
+                    : 2;
+
+                _stringKey = string.Create(
+                    digits,
+                    (key, digits),
+                    (span, tuple) =>
+                    {
+                        var (key, digits) = tuple;
+
+                        span[0] = '_';
+
+                        for (var i = digits - 1; i >= 1; i--)
+                        {
+                            span[i] = (char)('a' + key % 26);
+                            key /= 26;
+                        }
+                    });
+                #endif
             }
         }
 
@@ -32,7 +54,7 @@ namespace ExRam.Gremlinq.Core
 
         public static implicit operator string(BindingKey key) => key._stringKey is { } stringKey
             ? stringKey
-            : throw new ArgumentException();
+            : throw new ArgumentException(null, nameof(key));
 
         public override string ToString() => _stringKey is { } stringKey
             ? stringKey