```
| Method |      Mean |    Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |----------:|---------:|----------:|-------:|------:|------:|----------:|
|    Old | 101.32 ns | 3.566 ns | 10.514 ns | 0.0688 |     - |     - |     288 B |
|    New |  47.82 ns | 0.991 ns |  1.060 ns | 0.0095 |     - |     - |      40 B |